Old HTML renderer replaced with the new one.
